refactor(InvoiceDetailsDrawer): move the component to its own directory
