+++ src/components/drawers/Transactions/InvoiceDetailsDrawer/InvoiceDetailsDrawer.cy.tsx
