+++ src/pages/transacciones/factura/detail/index.tsx
@@ -8,7 +8,7 @@ import InvoiceForm from '~components/forms/Invoice'
 import EditToolbar from '~components/table/headers/toolbar'
 import AllDetailsTable from '~components/parameters/invoices/AllDetailsTable'
 import { useGetAllInvoiceDetailsQuery } from '~redux/api/bca-backend/transacciones/invoiceDetailsSlice'
-import InvoiceDetailsDrawer from '~components/drawers/Transactions/InvoiceDetailsDrawer'
+import InvoiceDetailsDrawer from '~/components/drawers/Transactions/InvoiceDetailsDrawer/InvoiceDetailsDrawer'
 
 export default function IndividualInvoice() {
     const [open, setOpen] = useState<boolean>(false)
